OpenDingux: GKD350h support
Start and Select are swapped in this version because Start + D-Pad controls backlight on the GKD350h.

There is nothing to map mouse emulation to in this version because of this but it isn't necessary to play the game.

This version looks uglier than on RG350 because it uses software scaling beacuse the IPU on the GKD350h doesn't work.